feat(mock-data): expand mock data generation to Collaborator and AdultStudent
Add mock data generators and factories for Collaborator and AdultStudent
entity types, extending the existing Employee-only implementation to
cover all currently available creation use cases.

Changes:
- Fix Long/Integer type mismatch in LoadEmployeeMockDataUseCase
- Extract shared person-level generation to PersonDataGenerator
- Refactor EmployeeDataGenerator to delegate to PersonDataGenerator
- Add CollaboratorDataGenerator, CollaboratorFactory,
  LoadCollaboratorMockDataUseCase
- Add AdultStudentDataGenerator, AdultStudentFactory,
  LoadAdultStudentMockDataUseCase
- Update PeopleDataLoaderConfiguration with Collaborator and
  AdultStudent DataLoader/DataCleanUp beans
- Update LoadPeopleMockDataUseCase to orchestrate all three entity
  types with FK-safe cleanup ordering
- Add CustomerAuth cleanup for AdultStudent teardown
- Remove unused SerializationFailureException
- Remove unused Tomcat ApplicationContext import from
  EmployeeDataGenerator

+++ mock-data-system/src/main/java/com/akademiaplus/users/usecases/LoadEmployeeMockDataUseCase.java
@@ -16,10 +16,10 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 @Service
-public class LoadEmployeeMockDataUseCase extends AbstractMockDataUseCase<EmployeeCreationRequestDTO, EmployeeDataModel, Integer> {
+public class LoadEmployeeMockDataUseCase extends AbstractMockDataUseCase<EmployeeCreationRequestDTO, EmployeeDataModel, Long> {
 
-    public LoadEmployeeMockDataUseCase(DataLoader<EmployeeCreationRequestDTO, EmployeeDataModel, Integer> dataLoader,
-                                       @Qualifier("employeeDataCleanUp") DataCleanUp<EmployeeDataModel, Integer> dataCleanup) {
+    public LoadEmployeeMockDataUseCase(DataLoader<EmployeeCreationRequestDTO, EmployeeDataModel, Long> dataLoader,
+                                       @Qualifier("employeeDataCleanUp") DataCleanUp<EmployeeDataModel, Long> dataCleanup) {
         super(dataLoader, dataCleanup);
     }
 }
